+++ CHANGELOG.md
@@ -9,6 +9,15 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 - `IEventSubscription` interface for any operation with Events
 
 
+## [1.2.1] - 2023-07-13
+### Added
+- Added `DateValidatorTag` to Complex Tags
+- Added `OwnerName` property to SignNowDocument model
+
+### Fixed
+- Fixed Json property name for attachment complex tag
+
+
 ## [1.2.0] - 2023-07-03
 ### Added
 - Added Readme and License files to Nuget package
@@ -186,8 +195,9 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 [create freeform invite]: https://github.com/signnow/SignNow.NET/blob/develop/README.md#create-freeform-invite
 
 <!-- Links to compare changes from previous version vs new version -->
-[Unreleased]: https://github.com/signnow/SignNow.NET/compare/1.2.0...HEAD
-[1.1.1]: https://github.com/signnow/SignNow.NET/compare/1.1.1...1.2.0
+[Unreleased]: https://github.com/signnow/SignNow.NET/compare/1.2.1...HEAD
+[1.2.1]: https://github.com/signnow/SignNow.NET/compare/1.2.0...1.2.1
+[1.2.0]: https://github.com/signnow/SignNow.NET/compare/1.1.1...1.2.0
 [1.1.1]: https://github.com/signnow/SignNow.NET/compare/1.1.0...1.1.1
 [1.1.0]: https://github.com/signnow/SignNow.NET/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/signnow/SignNow.NET/compare/0.9.0...1.0.0